Moving JaggedArray to its own file
