+++ armi/bookkeeping/db/database3.py
@@ -57,6 +57,7 @@ import h5py
 import numpy
 
 from armi import context, getApp, meta, runLog, settings
+from armi.bookkeeping.db.jaggedArray import JaggedArray
 from armi.bookkeeping.db.layout import (
     DB_VERSION,
     Layout,
@@ -92,168 +93,6 @@ def getH5GroupName(cycle: int, timeNode: int, statePointName: str = None) -> str
     simulated run.
     """
     return "c{:0>2}n{:0>2}{}".format(cycle, timeNode, statePointName or "")
-
-
-class JaggedArray:
-    def __init__(self, jaggedData, paramName):
-        """
-        Take a list of numpy arrays or lists and flatten them into a single 1D array.
-
-        This implementation can preserve the structure of a multi-dimensional numpy array
-        by storing the dimensions in self.shapes and then re-populating a numpy array of
-        that shape from the flattened 1D array. However, it can only preserve one layer of
-        jaggedness in a list of lists (or other iterables). For example, a list of tuples
-        with varying lengths can be flattened and reconstituted exactly. But, if a list of
-        lists of tuples is passed in, the tuples in that final layer of nesting will all be
-        flattened to a single 1D numpy array after a round trip. No structure is retained
-        from nested lists of lists or lists of tuples.
-
-        Parameters
-        ----------
-        jaggedData: list of numpy.ndarray
-            A list of numpy arrays (or lists) to be flattened into a single array
-        paramName: str
-            The name of the parameter represented by this data
-        """
-        offset = 0
-        flattenedArray = []
-        offsets = []
-        shapes = []
-        for array in jaggedData:
-            offsets.append(offset)
-            if isinstance(array, numpy.ndarray):
-                shapes.append(array.shape)
-                offset += array.size
-                flattenedArray.extend(array.flatten())
-            elif isinstance(array, (int, float, list, tuple)):
-                try:
-                    numpyArray = numpy.array(array)
-                    shapes.append(numpyArray.shape)
-                    offset += numpyArray.size
-                    flattenedArray.extend(numpyArray.flatten())
-                except:  # noqa: E722
-                    # numpy might fail if it's jagged
-                    flattenedList = self.flatten(array)
-                    shapes.append(
-                        len(flattenedList),
-                    )
-                    offset += len(flattenedList)
-                    flattenedArray.extend(flattenedList)
-            elif array is None:
-                flattenedArray.append(None)
-                offset += 1
-        self.flattenedArray = numpy.array(flattenedArray)
-        self.offsets = numpy.array(offsets)
-        self.shapes = numpy.array(shapes)
-        self.nones = numpy.array(
-            [i for i, val in enumerate(flattenedArray) if val is None]
-        )
-        self.dtype = self.flattenedArray.dtype
-        self.paramName = paramName
-
-    def __iter__(self):
-        """Iterate over the unpacked list."""
-        return iter(self.unpack())
-
-    def __contains__(self, other):
-        return other in self.flattenedArray
-
-    @staticmethod
-    def flatten(x):
-        if isinstance(x, list):
-            if len(x) == 0:
-                return []
-            first, rest = x[0], x[1:]
-            return JaggedArray.flatten(first) + JaggedArray.flatten(rest)
-        else:
-            return [x]
-
-    @classmethod
-    def fromH5(cls, data, offsets, shapes, nones, dtype, paramName):
-        """
-        Create a JaggedArray instance from an HDF5 dataset.
-
-        The JaggedArray is stored in HDF5 as a flat 1D array with accompanying
-        attributes of "offsets" and "shapes" to define how to reconstitute the
-        original data.
-
-        Parameters
-        ----------
-        data: numpy.ndarray
-            A flattened 1D numpy array read in from an HDF5 file
-        offsets: numpy.ndarray
-            Offset indices for the zeroth element of each constituent array
-        shapes: numpy.ndarray
-            The shape of each constituent array
-        nones: numpy.ndarray
-            The location of Nones
-        dtype: numpy.dtype
-            The data type for the array
-        paramName: str
-            The name of the parameter represented by this data
-
-        Returns
-        -------
-        obj: JaggedArray An instance of JaggedArray populated with the input data
-        """
-        obj = cls([], paramName)
-        obj.flattenedArray = numpy.array(data)
-        obj.offsets = numpy.array(offsets)
-        obj.shapes = numpy.array(shapes)
-        obj.nones = numpy.array(nones)
-        obj.dtype = dtype
-        obj.paramName = paramName
-        return obj
-
-    def tolist(self):
-        """Alias for unpack() to make this class respond like a numpy.ndarray."""
-        return self.unpack()
-
-    def unpack(self):
-        """
-        Unpack a JaggedArray object into a list of arrays.
-
-        Returns
-        -------
-        unpackedJaggedData: list of numpy.ndarray
-            List of numpy arrays with varying dimensions (i.e., jagged arrays)
-        """
-        unpackedJaggedData: List[Optional[numpy.ndarray]] = []
-        for offset, shape in zip(self.offsets, self.shapes):
-            unpackedJaggedData.append(
-                numpy.ndarray(
-                    shape, dtype=self.dtype, buffer=self.flattenedArray[offset:]
-                )
-            )
-        return self.replaceNonsenseWithNones(unpackedJaggedData)
-
-    def replaceNonsenseWithNones(self, unpackedData):
-        """Replace None markers in the unpacked array with an actual None."""
-        unpackedObject = []
-        for i, data in enumerate(unpackedData):
-            if numpy.issubdtype(self.dtype, numpy.floating):
-                isNone = numpy.isnan(data)
-            elif numpy.issubdtype(self.dtype, numpy.integer):
-                isNone = data == numpy.iinfo(self.dtype).min + 2
-            elif numpy.issubdtype(self.dtype, numpy.str_):
-                isNone = data == "<!None!>"
-            elif numpy.issubdtype(self.dtype, numpy.object_):
-                # object type means Nones are probably already present
-                isNone = data == None  # noqa: E711
-            else:
-                raise TypeError(
-                    "Unable to resolve values that should be None for `{}`. dtype = `{}`".format(
-                        self.paramName, self.dtype
-                    )
-                )
-
-            if data.ndim > 1 and isNone.all():
-                replacedData = None
-            else:
-                replacedData = numpy.array(data, dtype=numpy.dtype("O"))
-                replacedData[isNone] = None
-            unpackedObject.append(replacedData)
-        return unpackedObject
 
 
 class Database3:
